added missed deps and some refactoring
