+++ frontend/package.json
@@ -59,6 +59,7 @@
     "jest-environment-jsdom": "^29.5.0",
     "lint-staged": "^13.1.2",
     "postcss": "^8.4.23",
+    "postcss-nested": "^6.0.1",
     "prettier": "^2.8.7",
     "prettier-plugin-tailwindcss": "^0.2.6",
     "start-server-and-test": "^2.0.0",
