+++ frontend/postcss.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   plugins: {
+    'tailwindcss/nesting': 'postcss-nested',
     tailwindcss: {},
     autoprefixer: {},
   },
